Add switch to verify repo file paths
- Add new switch -verifyLocalRepo to check filenames and get their
correct case from a git repo.

This commit is a fix for a case-sensitivity issue:

Github servers expect case sensitive file paths, but we are relying on
the file paths from the PDB which may not be the same case as in the
repo. Without verifying the paths in the repo it's possible the source
server file path information written to the PDB may be incorrect.

'Github servers expect case sensitive file paths'
https://github.com/Haemoglobin/GitHub-Source-Indexer/issues/4